+++ aurora-backend/src/main/java/com/github/gr/aurora_bookstore/model/entity/OrderItem.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -30,8 +32,8 @@ public class OrderItem {
     private Integer quantity;
 
     @Column(name = "item_price", nullable = false)
-    private Integer itemPrice;
+    private BigDecimal itemPrice;
 
     @Column(name = "total_price", nullable = false)
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
 }
